refactor : 랜덤으로 움직이는 책임을 Car 로 위임
기존에 랜덤으로 움직이는 작업을 Cars 에서 해주었는데 움직임에 대한 책임은 Car 가 가지고 있는게 맞다고 생각하여 리팩터링
