+++ src/main/java/racingcar/domain/Car.java
@@ -1,6 +1,8 @@
 package racingcar.domain;
 
 
+import racingcar.Utils;
+
 public class Car {
     public static final String CAR_POSITION_SIGN = "-";
     public static final int START_POSITION = 0;
@@ -18,8 +20,10 @@ public class Car {
         return this.position.getAmount() == position;
     }
 
-    public void moveForward() {
-        position.increment(speed);
+    public void moveRandomly() {
+        if (Utils.isTrueWithProbability60Percent()) {
+            position.increment(speed);
+        }
     }
 
     public String getName() {
